+++ src/main/java/nextstep/courses/domain/Session.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.code.SessionType;
 import nextstep.payments.domain.Payment;
 
 public interface Session {
