+++ src/main/java/nextstep/courses/domain/Width.java
@@ -1,6 +1,6 @@
 package nextstep.courses.domain;
 
-import nextstep.courses.SizeUnderException;
+import nextstep.courses.exception.SizeUnderException;
 
 public class Width {
