+++ src/main/java/nextstep/courses/domain/code/SessionType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.code;
 
 public enum SessionType {
