+++ src/main/java/nextstep/courses/exception/NotExtensionTypeException.java
@@ -1,4 +1,4 @@
-package nextstep.courses;
+package nextstep.courses.exception;
 
 public class NotExtensionTypeException extends IllegalArgumentException {
     private static final long serialVersionUID = 1L;
