+++ src/test/java/nextstep/courses/domain/AmountTest.java
@@ -1,6 +1,6 @@
 package nextstep.courses.domain;
 
-import nextstep.courses.NotPositiveException;
+import nextstep.courses.exception.NotPositiveException;
 import nextstep.payments.domain.Payment;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
